Ultra otimização das classes de usuários com base em um diagrama

+++ app/adms/Controllers/usuarios/UsuariosAbstract.php
@@ -7,13 +7,41 @@ use App\adms\Helpers\PHPMailerHelper;
 use App\adms\Models\Repositories\TokenRepository;
 use App\adms\Models\Repositories\UsuariosRepository;
 use App\adms\Models\Services\DbConnectionClient;
+use App\adms\Views\Services\LoadViewService;
 
-/** Define funções universais para as classes de usuários. Tem o objetivo de ser herdado. Instancia o repositório automaticamente. */
-abstract class UsuariosAbstract 
+/** 
+ * Define funções universais para as classes de usuários. 
+ * Tem o objetivo de ser herdado. 
+ * Instancia o repositório automaticamente.
+ */
+abstract class UsuariosAbstract
 {
-  /** @var protected $repo O repositório de usuários  */
+  /** @var array $data Contém as informações da VIEW */
+  protected array $data = [
+    "title" => "Usuários",
+    "css" => ["public/adms/css/usuarios.css"],
+    "js" => []
+  ];
+
+  /** @var int $id O ID do usuário */
+  public int $id = 0;
+
+  /** @var string $nome O nome do usuário */
+  public string $nome = "";
+
+  /** @var string $email O e-mail do usuário */
+  public string $email = "";
+
+  /** @var int $statusId O ID do status do usuário. 1 = aguardando confirmação; 2 = Desativado; 3 = Ativado */
+  public int $statusId;
+
+  /** @var int $nivId O ID do nível de acesso */
+  public int $nivId;
+
+  /** @var protected $repo O repositório de usuários */
   protected $repo;
-  /** @var protected $tokenRepo O repositório de TOKENs  */
+
+  /** @var protected $tokenRepo O repositório de TOKENs */
   protected $tokenRepo;
 
   /** Conecta com o banco de dados do cliente depois inicia o repositório de usuários e de tokens */
@@ -24,15 +52,70 @@ abstract class UsuariosAbstract
     $this->tokenRepo = new TokenRepository($conn->conexao);
   }
 
+  protected function setData(array $data): void
+  {
+    $this->data = array_merge($this->data, $data);
+  }
+
+  protected function getData(): array
+  {
+    return $this->data;
+  }
+
+  protected function render(string $viewPath): void
+  {
+    $loadView = new LoadViewService($viewPath, $this->getData());
+    $loadView->loadView();
+  }
+
+  /** Recupera os dados do usuário e já trata os erros e faz o direcionamento caso falhe. */
+  public function setInfoById(int $usuarioId):void
+  {
+    // Recupera os dados do usuário
+    $this->id = (int)$usuarioId;
+    $usuarioArray = $this->repo->selecionar($this->id);
+
+    // Verifica se houve erro
+    if (($usuarioArray === false) || (empty($usuarioArray))) {
+      GenerateLog::generateLog("error", "A consulta ao repositório retornou falso ou vazio", ["id" => $usuarioId]);
+
+      // Prepara o setWarning
+      $_SESSION["alerta"] = [
+        "O usuário não existe!",
+        "O usuário não existe ou foi excluído."
+      ];
+
+      $this->redirect();
+    }
+
+    // Simplifica o array
+    $usuario = $usuarioArray[0];
+
+    // Seta os parâmetros do usuário
+    $this->nome = $usuario["u_nome"];
+    $this->email = $usuario["u_email"];
+    $this->nivId = $usuario["niv_id"];
+    $this->statusId = $usuario["us_id"];
+
+    // Seta como array também para passar para a VIEW se necessário
+    $this->data["usuario"] = $usuario;
+  }
+
   /**
    * Armazena uma foto de perfil de um usuário. Não funciona se o $_FILES["foto"] não estiver setado.
    * 
-   * @param int $usuarioId O ID do usuário
+   * Precisa dos parâmetros setados: id
    * 
-   * @return bool Se o upload falhou ou não
+   * @return bool Se o upload funcionou ou não
    */
-  public function armazenarFoto(int $usuarioId) :bool
+  public function armazenarFoto(): bool
   {
+    // Verifica se o id está setado
+    if ($this->id === 0){
+      GenerateLog::generateLog("error", "O ID do usuário não está setado.", ["metodo" => "armazenarFoto()"]);
+      return false;
+    }
+
     ini_set('file_uploads', '1');
 
     $extensoes_permitidas = [
@@ -44,19 +127,19 @@ abstract class UsuariosAbstract
     $tipo = mime_content_type($_FILES['foto']['tmp_name']);
 
     // Verifica se o tipo existe
-    if (!array_key_exists($tipo, $extensoes_permitidas)){
-      GenerateLog::generateLog("error", "Formato de arquivo inválido na tentativa de upload para foto de perfil.", ["tipo" => $tipo, "usuario_id" => $usuarioId, "filearray" => $_FILES["foto"]]);
+    if (!array_key_exists($tipo, $extensoes_permitidas)) {
+      GenerateLog::generateLog("error", "Formato de arquivo inválido na tentativa de upload para foto de perfil.", ["tipo" => $tipo, "usuario_id" => $this->id, "filearray" => $_FILES["foto"]]);
       return false;
     }
 
-    $arquivoNome = $usuarioId.$extensoes_permitidas[$tipo];
-    $caminho = "files/uploads/{$_SESSION['servidor_id']}/fotos-perfil/";  
+    $arquivoNome = $this->id . $extensoes_permitidas[$tipo];
+    $caminho = "files/uploads/{$_SESSION['servidor_id']}/fotos-perfil/";
 
     // Tenta criar o diretório se não existir
-    if (!is_dir($caminho)){
+    if (!is_dir($caminho)) {
       $novaPasta = mkdir($caminho, 0777, true);
       if (!$novaPasta) {
-        GenerateLog::generateLog("error", "Não foi possível criar a pasta.", ["pasta" => $caminho, "tipo" => $tipo, "usuario_id" => $usuarioId, "filearray" => $_FILES["foto"]]);
+        GenerateLog::generateLog("error", "Não foi possível criar a pasta.", ["pasta" => $caminho, "tipo" => $tipo, "usuario_id" => $this->id, "filearray" => $_FILES["foto"]]);
         return false;
       }
     }
@@ -65,23 +148,23 @@ abstract class UsuariosAbstract
     $imagem = getimagesize($tmp);
 
     // Verifica se é uma imagem
-    if (!$imagem){
-      GenerateLog::generateLog("error", "O arquivo não é uma imagem.", ["tipo" => $tipo, "usuario_id" => $usuarioId, "filearray" => $_FILES["foto"]]);
+    if (!$imagem) {
+      GenerateLog::generateLog("error", "O arquivo não é uma imagem.", ["tipo" => $tipo, "usuario_id" => $this->id, "filearray" => $_FILES["foto"]]);
       return false;
     }
 
-    $final = $caminho.$arquivoNome;
+    $final = $caminho . $arquivoNome;
 
     // Verifica se o arquivo existe
-    if (file_exists($final)){
-      GenerateLog::generateLog("error", "O arquivo já existe. Fique atento a IDs duplicados.", ["tipo" => $tipo, "usuario_id" => $usuarioId, "filearray" => $_FILES["foto"]]);
+    if (file_exists($final)) {
+      GenerateLog::generateLog("error", "O arquivo já existe. Fique atento a IDs duplicados.", ["tipo" => $tipo, "usuario_id" => $this->id, "filearray" => $_FILES["foto"]]);
       return false;
     }
 
     // Tenta criar os uploads
-    if (move_uploaded_file($tmp, $final)){
+    if (move_uploaded_file($tmp, $final)) {
       // Se o usuário for o próprio, adiciona na SESSION
-      if ($_SESSION["usuario_id"] === $usuarioId)
+      if ($_SESSION["usuario_id"] === $this->id)
         $_SESSION["foto_perfil"] = $final;
 
       // Atualiza o banco de dados
@@ -89,48 +172,58 @@ abstract class UsuariosAbstract
       $params = [
         ":foto_perfil" => $tipoFormatado
       ];
-      $this->repo->updateUsuario($params, $usuarioId);
+      $this->repo->updateUsuario($params, $this->id);
       return true;
     }
 
-    GenerateLog::generateLog("error", "Não pôde armazenar por algum motivo.", ["tipo" => $tipo, "usuario_id" => $usuarioId, "filearray" => $_FILES["foto"], "caminho" => $caminho, "final" => $final]);
+    GenerateLog::generateLog("error", "Não pôde armazenar por algum motivo.", ["tipo" => $tipo, "usuario_id" => $this->id, "filearray" => $_FILES["foto"], "caminho" => $caminho, "final" => $final]);
     return false;
   }
 
   /**
    * Apaga o arquivo da foto em uploads
    * 
-   * @param int $usuarioId O ID do usuário
+   * Precisa setar: id
    * 
    * @return bool Se apagou ou não
    */
-  public function apagarFoto(int $usuarioId):bool
+  public function apagarFoto(): bool
   {
-    $caminho = "files/uploads/{$_SESSION['servidor_id']}/fotos-perfil/$usuarioId";
+    // Verifica se o id está setado
+    if ($this->id === 0){
+      GenerateLog::generateLog("error", "O ID do usuário não está setado.", ["metodo" => "apagarFoto()"]);
+      return false;
+    }
 
-    $extensoes_permitidas = ['.jpg','.png','.jpeg'];
+    $caminho = "files/uploads/{$_SESSION['servidor_id']}/fotos-perfil/{$this->id}";
+
+    $extensoes_permitidas = ['.jpg', '.png', '.jpeg'];
 
     $ok = false;
 
     // Verica o arquivo para cada extensão
-    foreach ($extensoes_permitidas as $tipo){
-      $arquivo = $caminho.$tipo;
+    foreach ($extensoes_permitidas as $tipo) {
+      $arquivo = $caminho . $tipo;
 
       // Verifique se o arquivo existe
-      if (!empty($arquivo) && file_exists($arquivo)){
+      if (!empty($arquivo) && file_exists($arquivo)) {
         unlink($arquivo);
         $ok = true;
       }
     }
 
     if ($ok === false)
-      GenerateLog::generateLog("error", "A foto não foi apagada.", ["id" => $usuarioId]);
+      GenerateLog::generateLog("error", "A foto não foi apagada.", ["id" => $this->id]);
     else {
       // Atualiza o banco de dados
       $params = [
         ":foto_perfil" => null
       ];
-      $this->repo->updateUsuario($params, $usuarioId);
+      $this->repo->updateUsuario($params, $this->id);
+
+      // Remove da $_SESSION caso seja o mesmo usuário
+      if ($this->id === $_SESSION["usuario_id"])
+        $_SESSION["foto_perfil"] = null;
     }
 
     return $ok;
@@ -139,43 +232,51 @@ abstract class UsuariosAbstract
   /**
    * Troca a foto por uma nova, chamando a função apagarFoto e armazenarFoto
    * 
-   * @param int $usuarioId O ID do usuário
+   * Precisa setar: id
    * 
    * @return bool Se funcionou ou não
    */
-  public function trocarFoto(int $usuarioId):bool
+  public function trocarFoto(): bool
   {
-    $apagar = $this->apagarFoto($usuarioId);
+    $apagar = $this->apagarFoto();
 
     if ($apagar === false)
       return false;
 
-    return $this->armazenarFoto($usuarioId);
+    return $this->armazenarFoto();
   }
 
-    /**
+  /**
    * Enviar um email de confirmação
    * 
-   * @param int $usuarioId O ID do usuário.
-   * @param string $usuarioNome O nome do usuário.
-   * @param string $usuarioEmail O email do usuário.
+   * Precisa dos parâmetros setados: id, nome, email
    * 
-   * @return bool Se falhou ou não
+   * @return bool Se funcionou ou não
    */
-  public function emailConfirmacao(int $usuarioId, string $usuarioNome, string $usuarioEmail):bool
-  {    
+  public function emailConfirmacao(): bool
+  {
+    // Verifica se os parâmetros estão devidamente setados
+    if (($this->id === 0) || ($this->nome === "") || ($this->email === "")) {
+      GenerateLog::generateLog(
+        "error",
+        "O método emailConfirmacao() foi chamado, mas os parâmetros do usuário não estavam setados corretamente.",
+        ["id" => $this->id, "nome" => $this->nome, "email" => $this->nome]
+      );
+      return false;
+    }
+
     // Cria devidamente o TOKEN
-    $token = $this->tokenRepo->armazenarToken("sistema", "confirmar_email_senha", $usuarioId);
+    $token = $this->tokenRepo->armazenarToken("sistema", "confirmar_email_senha", $this->id);
 
     // Cria o email de envio
     $mail = new PHPMailerHelper;
 
-    $mail->destinatarios([$usuarioEmail]);
+    $mail->destinatarios([$this->email]);
 
     $mail->imagens([["caminho" => "public/adms/img/logo.png", "nome" => "logo"]]);
 
     $params = [
-      "[NOME]" => $usuarioNome,
+      "[NOME]" => $this->nome,
       "[SERVIDOR_ID]" => $_SESSION["servidor_id"],
       "[TOKEN]" => $token
     ];
@@ -253,5 +354,14 @@ abstract class UsuariosAbstract
     $title = "RD Mind | Confirme seu E-mail e Crie sua Senha";
     $mail->setarConteudo($title, $body);
     return $mail->enviar();
+  }
+
+  /**
+   * Redireciona de volta para listar usuário.
+   */
+  public function redirect(): void
+  {
+    header("Location: {$_ENV['HOST_BASE']}listar-usuarios");
+    exit;
   }
 }
